feat: start implementing ui
