+++ .gitignore copy
@@ -1,5 +1,3 @@
-/backend/target/
-
 # Logs
 logs
 *.log
