+++ backend/Cargo.toml
@@ -7,12 +7,13 @@ edition = "2021"
 serialport = "4.7.1"
 sqlite = "0.37.0"
 chrono = "0.4.41"
-
 num_enum = "0.7.3"
 thiserror = "2.0.12"
-
 regex = "1.11.1"
 evalexpr = "12.0.2"
-
 serde_json = "1.0.140"
 serde = { version = "^1.0.190", features = ["derive"] }
+tauri = { version = "1", features = ["shell-open"] }
+
+[build-dependencies]
+tauri-build = { version = "1", features = [] }
