+++ backend/src/pid/diagnostics.rs
